Exception handling - delete

+++ src/main/java/com/swprojects/webservices/services/UserService.java
@@ -4,11 +4,16 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.swprojects.webservices.entities.User;
 import com.swprojects.webservices.repositories.UserRepository;
+import com.swprojects.webservices.services.exceptions.DatabaseException;
 import com.swprojects.webservices.services.exceptions.ResourceNotFoundException;
+
+import jakarta.transaction.Transactional;
 
 @Service
 public class UserService {
@@ -29,8 +34,33 @@ public class UserService {
 		return userRepository.save(obj);
 	}
 
+	/*public void excluir(Long id) {
+
+		try {
+			userRepository.deleteById(id);
+
+		} catch (EmptyResultDataAccessException e) {
+			throw new ResourceNotFoundException(id);
+
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
+	}*/
+
+	@Transactional
 	public void delete(Long id) {
-		userRepository.deleteById(id);
+		if (!userRepository.existsById(id)) {
+			throw new ResourceNotFoundException(id);
+		}
+
+		try {
+			userRepository.deleteById(id);
+		} catch (EmptyResultDataAccessException e) {
+			throw new ResourceNotFoundException(id);
+
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 
 	public User update(Long id, User obj) {
